backends/stable_diffusion/requirements.txt
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
Pillow==9.4.0
Flask==3.1.0
